TwiliteTimeline

TwiliteTimeline.Authentictor

TwiliteTimeline.ResourceParams

TwiliteTimeline.collect_tweets

TwiliteTimeline.params_to_dict

TwiliteTimeline.Authentictor — Type

Authentictor (Datatype:TwiliteTimeline)

Authenticator is defined datatype to create an object which will be used to authenticate the user when interacting with the twitter API. Someone must generate the Consumer Key, Consumer Secret, OAuth Token and OAuth Secret using their own twitter developer account then using those credentionas create a value of the type Authenticator.

Example

# Here Twitter OAuth token and secret is stored as environment varialbe whith name
# CKEY, CSEC, OTOK, OSEC
julia> creds =  Authentictor(ENV["CKEY"], ENV["CSEC"], ENV["OTOK"], ENV["OSEC"])

TwiliteTimeline.ResourceParams — Type

ResourceParams(DataType:TwiliteTimeline)

ResourceParams is a defined datatype to store the Resource Parameter mentioned in the twitter API document which we have to send when qurying the API in the request body.

To be able to use twitter API the fild value shoud satisfy the following criteria:

When Providing valid count value the include_rts arguments must have to be 1; by default providing any valid count value of the include_rts will be set to 1 silintly

When providing constructor argument value for the boolean field such as trim_user, exclude_replies and include_rts the value mush have to be 1 or 0. By the default the values are set to the default value according to Twitter API doc.

Example

julia> resource_params = ResourceParams("realDonaldTrump", count=100)
ResourceParams("realDonaldTrump", nothing, nothing, 100, nothing, false, true)

julia> resource_params = ResourceParams("realDonaldTrump", count=100, exclude_replies=true)
ResourceParams("realDonaldTrump", nothing, nothing, 100, nothing, false, true)

TwiliteTimeline.collect_tweets — Method

collect_tweets(a::Authentictor, rp::ResourceParams)

Try to collect tweets based on given object of type Autenticator and ResourceParams. The collect_tweets is a wrapper around the oauth_request_resource method from OAuth package which is responsible for sending the get request to the statuses/user_timeline.

Example

# Here Twitter OAuth token and secret is stored as environment varialbe whith name
# CKEY, CSEC, OTOK, OSEC
julia> creds = Authentictor(ENV["CKEY"], ENV["CSEC"], ENV["OTOK"], ENV["OSEC"])
julia> resource_params = ResourceParams("realDonaldTrump", count=100)
julia> tweets = collect_tweets(creds, resource_params);

That should return 100 latest tweets from the profile @realDonaldTrump.

TwiliteTimeline.params_to_dict — Method

params_to_dict(rp::ResourceParams)

Is a internal function to convert the user provided Twitter API parameter into dictionary which oauth_request_resource from OAuth package expect. The method silently add include_rts paramater value to 1 when providing an valid count paramter.